# api/docs/swagger.yaml
definitions:
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  schedule_service.CreateTask:
    properties:
      deadline:
        type: string
      label:
        type: string
      lesson_id:
        type: string
      score:
        type: number
    type: object
...
  schedule_service.UpdateTask:
    properties:
      deadline:
        type: string
      id:
        type: string
      label:
        type: string
      lesson_id:
        type: string
      score:
        type: number
    type: object
...
  user_service.CreateSupportTeacher:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      ielts_attepms_count:
        type: integer
      ielts_score:
        type: number
      months_worked:
        type: integer
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
  user_service.UpdateSupportTeacher:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      ielts_attepms_count:
        type: integer
      ielts_score:
        type: number
      months_worked:
        type: integer
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CRM API
  version: "1.0"
paths:
...
  /api/v1/support-teacher:
    get:
      consumes:
      - application/json
      description: API for Get all support teachers
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all support teachers
      tags:
      - SUPPORT-TEACHERS
    post:
      consumes:
      - application/json
      description: API for creating support teacher
      parameters:
      - description: support-teacher
        in: body
        name: support-teacher
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateSupportTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a support teacher
      tags:
      - SUPPORT-TEACHERS
    put:
      consumes:
      - application/json
      description: API for update a support teacher
      parameters:
      - description: support-teacher
        in: body
        name: support-teacher
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateSupportTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a support teacher
      tags:
      - SUPPORT-TEACHERS
  /api/v1/support-teacher/{id}:
    delete:
      consumes:
      - application/json
      description: API for delete a support teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a support teacher
      tags:
      - SUPPORT-TEACHERS
    get:
      consumes:
      - application/json
      description: API for getting a support teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a student
      tags:
      - SUPPORT-TEACHERS
  /api/v1/task:
    post:
      consumes:
      - application/json
      description: API for creating task
      parameters:
      - description: schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/schedule_service.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a task
      tags:
      - TASKS
    put:
      consumes:
      - application/json
      description: API for update a task
      parameters:
      - description: schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/schedule_service.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a task
      tags:
      - TASKS
  /api/v1/task/{id}:
    delete:
      consumes:
      - application/json
      description: API for delete a task
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a task
      tags:
      - TASKS
    get:
      consumes:
      - application/json
      description: API for getting a task
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a task
      tags:
      - TASKS
  /api/v1/tasks:
    get:
      consumes:
      - application/json
      description: API for Get all tasks
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all tasks
      tags:
      - TASKS
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"